Fixed generating signals data (#1)

+++ SDR.Models/Configurations/RandomSignalDataProviderConfiguration.cs
@@ -2,10 +2,10 @@
 
 public class RandomSignalDataProviderConfiguration
 {
-    public byte FrequencyMin { get; set; }
-    public byte FrequencyMax { get; set; }
-    public sbyte StrengthMin { get; set; }
-    public sbyte StrengthMax { get; set; }
+    public float FrequencyMin { get; set; }
+    public float FrequencyMax { get; set; }
+    public float StrengthMin { get; set; }
+    public float StrengthMax { get; set; }
     public int Count { get; set; }
     public int Frequency { get; set; }
 }
