+++ SDR.Models/Signal.cs
@@ -2,9 +2,9 @@
 
 namespace SDR.Models;
 
-public readonly struct Signal(byte frequency, sbyte strength)
+public readonly struct Signal(float frequency, float strength)
 {
-    public byte Frequency { get; } = frequency;
-    public sbyte Strength { get; } = strength;
+    public float Frequency { get; } = frequency;
+    public float Strength { get; } = strength;
     public DateTime Time { get; } = DateTime.Now;
 }
